Create README.md
This is a futuristic technology project – “Augmented Reality + IOT Smart BULB”. My project is based on the combination of the Internet of things [IOT] and Augmented Reality technology.
In this project, I have controlled an Electric bulb over the Internet using a Virtual Button Control Panel. As seen in the video when I bring my device camera over the image target, the image target is detected and the Virtual Button Panel appears in the air, and using the ‘ON/OFF’ button present on the panel I can control (switch ON and OFF) my electric bulb.
In this project, I have used ESP32 Microcontroller, Relay channel, Blynk API, Unity, and Vuforia engines and an image target.